refactor: 출처목록 테스트 병합
